Make all Actions Async

+++ RehabRally.Web/Core/Models/PatientConclusion.cs
@@ -1,8 +1,8 @@
 ﻿namespace RehabRally.Web.Core.Models
 {
     public class PatientConclusion
-    {
-         public byte Id { get; set;}
+    { 
+        public int Id { get; set;}
         public string Conclusion { get; set; } = null!; 
         public string UserId { get; set; } = null!; 
     }
